Convert LoadBalance.actor.h to standard coroutines
